+++ www/js/survey/enketo/AddNoteButton.tsx
@@ -11,7 +11,7 @@ import React, { useEffect, useState, useContext } from 'react';
 import DiaryButton from '../../components/DiaryButton';
 import { useTranslation } from 'react-i18next';
 import { DateTime } from 'luxon';
-import LabelTabContext from '../../diary/LabelTabContext';
+import TimelineContext from '../../TimelineContext';
 import EnketoModal from './EnketoModal';
 import { displayErrorMsg, logDebug } from '../../plugin/logger';
 import { isTrip } from '../../types/diaryTypes';
@@ -24,7 +24,7 @@ type Props = {
 const AddNoteButton = ({ timelineEntry, notesConfig, storeKey }: Props) => {
   const { t, i18n } = useTranslation();
   const [displayLabel, setDisplayLabel] = useState('');
-  const { notesFor, addUserInputToEntry } = useContext(LabelTabContext);
+  const { notesFor, addUserInputToEntry } = useContext(TimelineContext);
 
   useEffect(() => {
     let newLabel: string;
